# Fix memoization of localizer data, add benchmark