Benchmark output

	go test -benchmem -run=^$ github.com/klauspost/lctime -bench ^(BenchmarkNewLocalizer)$

	goos: darwin
	goarch: amd64
	pkg: github.com/klauspost/lctime

	BEFORE:
	BenchmarkNewLocalizer-12    	   43866	     27375 ns/op	   46368 B/op	      94 allocs/op
	PASS
	ok  	github.com/klauspost/lctime	3.591s

	AFTER:
	BenchmarkNewLocalizer-12    	33426420	        35.8 ns/op	       0 B/op	       0 allocs/op
	PASS
	ok  	github.com/klauspost/lctime	2.929s